requirements.txt
insightface
onnxruntime
opencv-python
numpy
pillow
streamlit
